Codechange: add CopyOutDParam variant for StringParameters &&
