+++ src/strings_func.h
@@ -100,6 +100,7 @@ void SetDParamStr(size_t n, std::string &&str);
 
 void CopyInDParam(const std::span<const StringParameterBackup> backup);
 void CopyOutDParam(std::vector<StringParameterBackup> &backup, size_t num);
+void CopyOutDParam(std::vector<StringParameterBackup> &backup, class StringParameters &&params);
 bool HaveDParamChanged(const std::vector<StringParameterBackup> &backup);
 
 uint64_t GetDParam(size_t n);
